recommendation-service/requirements.txt
implicit
kafka-python
numpy
pandas
scipy
Flask